requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
arrow==1.2.3
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
beartype==0.15.0
beautifulsoup4==4.12.2
blessed==1.20.0
boto3==1.34.44
botocore==1.34.44
cachetools==5.3.1
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.1.1
chex==0.1.7
click==8.1.6
cmake==3.25.0
contourpy==1.1.0
croniter==1.4.1
cycler==0.11.0
dateutils==0.6.12
dbus-python==1.2.16
deepdiff==6.3.1
dill==0.3.7
dm-tree==0.1.8
docker-pycreds==0.4.0
easydict==1.10
editor==1.6.6
emlp==1.0.3
exceptiongroup==1.1.2
fastapi==0.100.1
filelock==3.9.0
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
fvcore==0.1.5.post20221221
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
graphviz==0.20.1
grpcio==1.56.2
h11==0.14.0
h5py==3.9.0
huggingface-hub==0.20.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
iniconfig==2.0.0
inquirer==3.1.3
iopath==0.1.10
itsdangerous==2.1.2
jax==0.4.24
jaxlib==0.4.24
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.4
lightning==2.0.6
lightning-cloud==0.5.37
lightning-utilities==0.9.0
lit==15.0.7
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.2
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
natsort==8.4.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
objax==1.8.0
opt-einsum==3.3.0
optax==0.1.9
ordered-set==4.1.0
packaging==23.1
pandas==2.0.3
parameterized==0.9.0
pathtools==0.1.2
Pillow==9.3.0
pluggy==1.3.0
plum-dispatch==2.1.1
portalocker==2.7.0
protobuf==4.23.4
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pydantic==2.0.3
pydantic_core==2.3.0
pyg-lib==0.4.0+pt22cu118
Pygments==2.15.1
PyJWT==2.8.0
pynauty==2.8.6
pyparsing==3.0.9
pyre-extensions==0.0.29
pytest==7.4.2
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.6
pytz==2023.3
PyYAML==6.0.1
readchar==4.0.5
regex==2023.6.3
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rich==13.5.2
rsa==4.9
runs==1.2.2
s3transfer==0.10.0
safetensors==0.4.2
scikit-learn==1.3.0
scipy==1.12.0
seaborn==0.12.2
sentry-sdk==1.29.2
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
starsessions==1.3.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
threadpoolctl==3.2.0
timm==0.9.5
tokenizers==0.19.1
tomli==2.0.1
toolz==0.12.0
torch==2.2.0+cu118
torch-tb-profiler==0.4.1
torch_cluster==1.6.3+pt22cu118
torch_geometric==2.5.0
torch_scatter==2.1.2+pt22cu118
torch_sparse==0.6.18+pt22cu118
torch_spline_conv==1.2.2+pt22cu118
torchaudio==2.2.0+cu118
torchcontrib==0.0.2
torchmetrics==1.0.2
torchvision==0.17.0+cu118
torchviz==0.0.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.40.1
triton==2.2.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.13
uvicorn==0.23.2
wandb==0.16.3
wcwidth==0.2.6
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.6
xformers==0.0.24+cu118
xmod==1.8.1
xxhash==3.3.0
yacs==0.1.8
yarl==1.9.2
zipp==3.16.2
